+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Release history
 0.3.2 (November 2013)
 +++++++++++++++++++++
 - fixed: typos, test structure
+- more dynamic importing of detector code
 
 0.3.1 (November 2013)
 +++++++++++++++++++++